fix: remove soft deletes

+++ database/migrations/2021_02_11_154933_create_subjects_table.php
@@ -13,8 +13,8 @@ class CreateSubjectsTable extends Migration
             $table->foreignId('user_id')->comment('subject teacher')->constrained();
             $table->foreignId('education_level_id')->constrained();
             $table->string('name');
-            $table->integer('max_marks');
-            $table->integer('min_passing_marks');
+            $table->tinyInteger('max_marks');
+            $table->tinyInteger('min_passing_marks');
             $table->timestamps();
         });
     }
